+++ backend/src/api/v1/tracks.py
@@ -1,5 +1,6 @@
 from fastapi import UploadFile, File, APIRouter, Depends, status
 
+from src.api.exceptions import NoRightsException
 from src.core.cruds import MediaRepository
 from src.schemas.auth import User
 
@@ -89,16 +90,21 @@ async def add_track(
     responses={status.HTTP_200_OK: {"model": SUpdateTrackPictureResponse}},
 )
 async def update_pic_tracks(
-    tracks_id: int,
+    track_id: int,
     file: UploadFile = File(...),
     user: User = Depends(get_current_user),
 ) -> SUpdateTrackPictureResponse:
+    track = await TracksRepository.find_one_by_id(id_=track_id)
+
+    if track.id != user.id:
+        raise NoRightsException()
+
     file_info = await unique_filename(file) if file else None
     file_url = await MediaRepository.upload_file("PICTURES", file_info, file)
 
     data = {"picture_url": file_url}
 
-    response = await TracksRepository.edit_one(tracks_id, data)
+    response = await TracksRepository.edit_one(track_id, data)
     return SUpdateTrackPictureResponse.from_db_model(model=response)
 
 
@@ -113,6 +119,11 @@ async def release_track(
     tracks_data: SReleaseTrackRequest,
     user: User = Depends(get_current_user)
 ) -> SReleaseTrackResponse:
+    track = await TracksRepository.find_one_by_id(id_=track_id)
+
+    if track.id != user.id:
+        raise NoRightsException()
+
     data = {
         "name": tracks_data.title,
         "description": tracks_data.description,
@@ -134,6 +145,11 @@ async def update_track(
     tracks_data: SUpdateTrackRequest,
     user: User = Depends(get_current_user)
 ) -> SUpdateTrackResponse:
+    track = await TracksRepository.find_one_by_id(id_=track_id)
+
+    if track.id != user.id:
+        raise NoRightsException()
+
     data = {
         "name": tracks_data.title,
         "description": tracks_data.description,
@@ -150,6 +166,14 @@ async def update_track(
     response_model=SDeleteTrackResponse,
     responses={status.HTTP_200_OK: {"model": SDeleteTrackResponse}},
 )
-async def delete_tracks(track_id: int):
+async def delete_tracks(
+        track_id: int,
+        user: User = Depends(get_current_user)
+):
+    track = await TracksRepository.find_one_by_id(id_=track_id)
+
+    if track.id != user.id:
+        raise NoRightsException()
+
     await TracksRepository.delete(id_=track_id)
     return SDeleteTrackResponse()
